requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
anthropic
Werkzeug==3.0.1
gunicorn==21.2.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9